[UT] Add IT framework for pipeline from kakfka to starrocks (#249)
