+++ src/test/java/com/starrocks/connector/flink/it/sink/StarRocksTableUtils.java
@@ -1,0 +1,73 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.starrocks.connector.flink.it.sink;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.StringJoiner;
+
+import static org.junit.Assert.assertArrayEquals;
+
+public class StarRocksTableUtils {
+
+    // Scan table, and returns a collection of rows
+    public static List<List<Object>> scanTable(Connection dbConnector, String db, String table) throws SQLException {
+        try (PreparedStatement statement = dbConnector.prepareStatement(String.format("SELECT * FROM `%s`.`%s`", db, table))) {
+            try (ResultSet resultSet = statement.executeQuery()) {
+                List<List<Object>> results = new ArrayList<>();
+                int numColumns = resultSet.getMetaData().getColumnCount();
+                while (resultSet.next()) {
+                    List<Object> row = new ArrayList<>();
+                    for (int i = 1; i <= numColumns; i++) {
+                        row.add(resultSet.getObject(i));
+                    }
+                    results.add(row);
+                }
+                return results;
+            }
+        }
+    }
+
+    public static void verifyResult(List<List<Object>> expected, List<List<Object>> actual) {
+        List<String> expectedRows = new ArrayList<>();
+        List<String> actualRows = new ArrayList<>();
+        for (List<Object> row : expected) {
+            StringJoiner joiner = new StringJoiner(",");
+            for (Object col : row) {
+                joiner.add(col == null ? "null" : col.toString());
+            }
+            expectedRows.add(joiner.toString());
+        }
+        expectedRows.sort(String::compareTo);
+
+        for (List<Object> row : actual) {
+            StringJoiner joiner = new StringJoiner(",");
+            for (Object col : row) {
+                joiner.add(col == null ? "null" : col.toString());
+            }
+            actualRows.add(joiner.toString());
+        }
+        actualRows.sort(String::compareTo);
+        assertArrayEquals(expectedRows.toArray(), actualRows.toArray());
+    }
+}
