+++ src/test/resources/data/debezium-upsert-and-delete.txt
@@ -1,0 +1,5 @@
+{"op":"c","ts_ms":1589355606100,"before":null,"after":{"c0":1,"c1":1.0,"c2":"row1"},"source":{"version":"1.1.1.Final","connector":"mysql","name":"dbserver1","ts_ms":0,"snapshot":"true","db":"inventory","table":"products","server_id":0,"gtid":null,"file":"mysql-bin.000003","pos":154,"row":0,"thread":null,"query":null},"transaction":null}
+{"op":"c","ts_ms":1589355606200,"before":null,"after":{"c0":2,"c1":2.0,"c2":"row2"},"source":{"version":"1.1.1.Final","connector":"mysql","name":"dbserver1","ts_ms":0,"snapshot":"true","db":"inventory","table":"products","server_id":0,"gtid":null,"file":"mysql-bin.000003","pos":154,"row":0,"thread":null,"query":null},"transaction":null}
+{"op":"c","ts_ms":1589355606300,"before":null,"after":{"c0":3,"c1":3.0,"c2":"row3"},"source":{"version":"1.1.1.Final","connector":"mysql","name":"dbserver1","ts_ms":0,"snapshot":"true","db":"inventory","table":"products","server_id":0,"gtid":null,"file":"mysql-bin.000003","pos":154,"row":0,"thread":null,"query":null},"transaction":null}
+{"op":"u","ts_ms":1589355606400,"before":{"c0":1,"c1":1.0,"c2":"row1"},"after":{"c0":1,"c1":2.0,"c2":"row1"},"source":{"version":"1.1.1.Final","connector":"mysql","name":"dbserver1","ts_ms":0,"snapshot":"true","db":"inventory","table":"products","server_id":0,"gtid":null,"file":"mysql-bin.000003","pos":154,"row":0,"thread":null,"query":null},"transaction":null}
+{"op":"d","ts_ms":1589355606500,"before":{"c0":3,"c1":3.0,"c2":"row3"},"after":null,"source":{"version":"1.1.1.Final","connector":"mysql","name":"dbserver1","ts_ms":0,"snapshot":"true","db":"inventory","table":"products","server_id":0,"gtid":null,"file":"mysql-bin.000003","pos":154,"row":0,"thread":null,"query":null},"transaction":null}
